backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
requests==2.31.0
django-filter==23.1
psycopg2-binary==2.9.3
Pillow